pyproject.toml
[project]
name = "llm-speedway"
version = "0.1.0"
description = "Benchmark harness for timing agentic coding tools"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-timeout>=2.3.0",
    "pyyaml>=6.0.0",
    "pexpect>=4.9.0",
    "psutil>=5.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["harness"]
